修复 flow 类型由 Long 改 Double 引发的编译错误

+++ springboot-backend/src/main/java/com/admin/controller/OpenApiController.java
@@ -66,7 +66,7 @@ public class OpenApiController extends BaseController {
             headerValue = buildSubscriptionHeader(
                     tunnelInfo.getOutFlow(),
                     tunnelInfo.getInFlow(),
-                    tunnelInfo.getFlow() * GIGA,
+                    (long)(tunnelInfo.getFlow() * GIGA),
                     tunnelInfo.getExpTime() / 1000
             );
         }
